complete user profile front and back and improve the not showing imag…

+++ resources/views/admin/user/index.blade.php
@@ -67,7 +67,7 @@
                                     class="p-1 lg:p-3 text-xs lg:text-sm h-full flex items-center justify-center text-gray-900 col-span-2">
                                     <div class="w-20 lg:w-full">
                                         <img class="max-w-[50px] max-h-[50px] mx-auto size-12 object-cover rounded-md"
-                                            src={{ isset($user->image) ? asset($user->image) : asset('assets/img/user.png') }}>
+                                            src={{ isset($user->mainImage) ? asset('storage/' . $user->mainImage) : asset('assets/img/user.png') }}>
                                     </div>
                                 </div>
                                 <div
@@ -81,15 +81,29 @@
                                 <div
                                     class="p-1 lg:p-3 text-xs lg:text-sm h-full flex items-center justify-center text-gray-900 text-center col-span-2">
                                     <span class="block w-20 max-h-15 lg:w-full overflow-auto" style="scrollbar-width: none">
-                                        @foreach ($user->roles as $role)
-                                            {{ $role['name'] }} -
+                                        @foreach ($user->roles as $index => $role)
+                                            @if ($index == count($user->roles) - 1)
+                                                @if ($role['name'] == 'admin')
+                                                    <span>ادمین</span>
+                                                @endif
+                                                @if ($role['name'] == 'general_user')
+                                                    <span>کاربر عمومی</span>
+                                                @endif
+                                            @else
+                                                @if ($role['name'] == 'admin')
+                                                    <span>ادمین - </span>
+                                                @endif
+                                                @if ($role['name'] == 'general_user')
+                                                    <span>کاربر عمومی - </span>
+                                                @endif
+                                            @endif
                                         @endforeach
                                     </span>
                                 </div>
                                 <div class="col-span-2 flex items-center">
                                     <ul class="w-[220px] lg:w-full text-sm rounded-sm p-1 grid grid-cols-3">
-                                        {{-- <li class="flex justify-center">
-                                            <span
+                                        <li class="flex justify-center">
+                                            <a href="{{ route('user.profile', [$user]) }}"
                                                 class="w-fit flex flex-row items-center justify-center bg-sky-500 hover:bg-sky-600 p-1 rounded-sm"
                                                 title="مشاهده">
                                                 <svg xmlns="http://www.w3.org/2000/svg" class="size-4"
@@ -97,8 +111,8 @@
                                                     <path fill="white"
                                                         d="M288 80c-65.2 0-118.8 29.6-159.9 67.7C89.6 183.5 63 226 49.4 256c13.6 30 40.2 72.5 78.6 108.3C169.2 402.4 222.8 432 288 432s118.8-29.6 159.9-67.7C486.4 328.5 513 286 526.6 256c-13.6-30-40.2-72.5-78.6-108.3C406.8 109.6 353.2 80 288 80zM95.4 112.6C142.5 68.8 207.2 32 288 32s145.5 36.8 192.6 80.6c46.8 43.5 78.1 95.4 93 131.1c3.3 7.9 3.3 16.7 0 24.6c-14.9 35.7-46.2 87.7-93 131.1C433.5 443.2 368.8 480 288 480s-145.5-36.8-192.6-80.6C48.6 356 17.3 304 2.5 268.3c-3.3-7.9-3.3-16.7 0-24.6C17.3 208 48.6 156 95.4 112.6zM288 336c44.2 0 80-35.8 80-80s-35.8-80-80-80c-.7 0-1.3 0-2 0c1.3 5.1 2 10.5 2 16c0 35.3-28.7 64-64 64c-5.5 0-10.9-.7-16-2c0 .7 0 1.3 0 2c0 44.2 35.8 80 80 80zm0-208a128 128 0 1 1 0 256 128 128 0 1 1 0-256z" />
                                                 </svg>
-                                            </span>
-                                        </li> --}}
+                                            </a>
+                                        </li>
                                         <li class="flex justify-center">
                                             <span onclick="editForm('open', {{ $user['id'] }})"
                                                 class="w-fit flex flex-row items-center justify-center bg-green-500 hover:bg-green-600 p-1 rounded-sm cursor-pointer"
@@ -110,18 +124,20 @@
                                                 </svg>
                                             </span>
                                         </li>
-                                        <li class="flex justify-center">
-                                            <span
-                                                onclick="deleteUser('open', {{ $user['id'] }}, '{{ $user['name'] }} {{ $user['family'] }}')"
-                                                class="w-fit flex flex-row items-center justify-center bg-red-500 hover:bg-red-600 p-1 rounded-sm cursor-pointer"
-                                                title="حذف">
-                                                <svg xmlns="http://www.w3.org/2000/svg" class="size-4"
-                                                    viewBox="0 0 448 512">
-                                                    <path fill="white"
-                                                        d="M170.5 51.6L151.5 80h145l-19-28.4c-1.5-2.2-4-3.6-6.7-3.6H177.1c-2.7 0-5.2 1.3-6.7 3.6zm147-26.6L354.2 80H368h48 8c13.3 0 24 10.7 24 24s-10.7 24-24 24h-8V432c0 44.2-35.8 80-80 80H112c-44.2 0-80-35.8-80-80V128H24c-13.3 0-24-10.7-24-24S10.7 80 24 80h8H80 93.8l36.7-55.1C140.9 9.4 158.4 0 177.1 0h93.7c18.7 0 36.2 9.4 46.6 24.9zM80 128V432c0 17.7 14.3 32 32 32H336c17.7 0 32-14.3 32-32V128H80zm80 64V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16z" />
-                                                </svg>
-                                            </span>
-                                        </li>
+                                        @if ($user->id != Auth::user()['id'])
+                                            <li class="flex justify-center">
+                                                <span
+                                                    onclick="deleteUser('open', {{ $user['id'] }}, '{{ $user['name'] }} {{ $user['family'] }}')"
+                                                    class="w-fit flex flex-row items-center justify-center bg-red-500 hover:bg-red-600 p-1 rounded-sm cursor-pointer"
+                                                    title="حذف">
+                                                    <svg xmlns="http://www.w3.org/2000/svg" class="size-4"
+                                                        viewBox="0 0 448 512">
+                                                        <path fill="white"
+                                                            d="M170.5 51.6L151.5 80h145l-19-28.4c-1.5-2.2-4-3.6-6.7-3.6H177.1c-2.7 0-5.2 1.3-6.7 3.6zm147-26.6L354.2 80H368h48 8c13.3 0 24 10.7 24 24s-10.7 24-24 24h-8V432c0 44.2-35.8 80-80 80H112c-44.2 0-80-35.8-80-80V128H24c-13.3 0-24-10.7-24-24S10.7 80 24 80h8H80 93.8l36.7-55.1C140.9 9.4 158.4 0 177.1 0h93.7c18.7 0 36.2 9.4 46.6 24.9zM80 128V432c0 17.7 14.3 32 32 32H336c17.7 0 32-14.3 32-32V128H80zm80 64V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16zm80 0V400c0 8.8-7.2 16-16 16s-16-7.2-16-16V192c0-8.8 7.2-16 16-16s16 7.2 16 16z" />
+                                                    </svg>
+                                                </span>
+                                            </li>
+                                        @endif
                                     </ul>
                                 </div>
                             </div>
